+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Nothing here yet...
 ### Added
 
 - Add options to hide background grid, transformed grid, and basis vectors
+- Fully respect display settings in visual definition widget
 
 ### Fixed